Updated bootstrap.sh for bind9 configuration
(1) Updated the designate.conf section to match the
designate.conf.sample section
(2) Replaced 2 separate designate.conf sections - 1 for bind and 1 for
powerdns with mostly 1 section for both.  Designate.conf now has all the
options specified for both bind and powerdns.  Based on which is enabled
service:central's backend_driver is updated towards the end of the file.
(3) v1 and v2 api's are enabled by default to make testing easier.  Also
the api_base_uri is set, so that the links in the responses work
correctly.
(4) I also added the current tlds that we have in cloud dns as tlds.txt.
This can be used with designate-manage (from a vagrant ssh window) to
import all the tlds as follows
designate-manage import-tlds
--input-file="/home/vagrant/designate/tlds.txt"